Address Copilot review comments: remove redundant long casts
Copilot comment:

[Testing] The `equalTo(AttributeKey<Long>, int)` overload already handles this value, so this cast is redundant under the Java test rules.

Copilot comment:

[Testing] The `equalTo(AttributeKey<Long>, int)` overload already handles this value, so this cast is redundant under the Java test rules.

Copilot comment:

[Testing] The `equalTo(AttributeKey<Long>, int)` overload already handles this value, so this cast is redundant under the Java test rules.

Copilot comment:

[Testing] The `equalTo(AttributeKey<Long>, int)` overload already handles this value, so this cast is redundant under the Java test rules.

Analysis: Each port expression already has type int. The attribute assertion API widens that value through its AttributeKey<Long> overload.

Upsides: The assertions follow the repository's test rules and keep the original expected values.

Downsides: No material downside identified.

+++ instrumentation/elasticsearch/elasticsearch-transport-common-5.0/testing/src/main/java/io/opentelemetry/javaagent/instrumentation/elasticsearch/transport/common/v5_0/AbstractElasticsearchTransportClientTest.java
@@ -122,7 +122,7 @@ public abstract class AbstractElasticsearchTransportClientTest
     List<AttributeAssertion> result = new ArrayList<>(asList(assertions));
     if (emitStableDatabaseSemconv()) {
       result.add(equalTo(SERVER_ADDRESS, getAddress()));
-      result.add(equalTo(SERVER_PORT, (long) getPort()));
+      result.add(equalTo(SERVER_PORT, getPort()));
     }
     return result;
   }
