+++ instrumentation/opensearch/opensearch-java-3.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/opensearch/v3_0/AbstractOpenSearchQueryTest.java
@@ -150,7 +150,7 @@ abstract class AbstractOpenSearchQueryTest {
     List<AttributeAssertion> result = new ArrayList<>(asList(assertions));
     if (emitStableDatabaseSemconv()) {
       result.add(equalTo(SERVER_ADDRESS, httpHost.getHost()));
-      result.add(equalTo(SERVER_PORT, (long) httpHost.getPort()));
+      result.add(equalTo(SERVER_PORT, httpHost.getPort()));
     }
     return result;
   }
